refactor: move away language services from `@volar/vue-typescript`
