+++ packages/typescript-vue-plugin/src/index.ts
@@ -2,7 +2,7 @@ import * as vue from '@volar/vue-typescript';
 import * as shared from '@volar/shared';
 import * as path from 'upath';
 import * as apis from './apis';
-import { createBasicRuntime } from '@volar/vue-typescript';
+import * as pug from '@volar/pug-language-service';
 
 const init: ts.server.PluginModuleFactory = (modules) => {
 	const { typescript: ts } = modules;
@@ -24,9 +24,54 @@ const init: ts.server.PluginModuleFactory = (modules) => {
 			};
 
 			const proxyHost = createProxyHost(ts, info);
-			const services = createBasicRuntime(undefined);
 			const vueCompilerOptions = proxyHost.host.getVueCompilationSettings?.() ?? {};
-			const tsRuntime = vue.createTypeScriptRuntime({ typescript: ts, ...services, vueCompilerOptions }, proxyHost.host, true);
+			const tsRuntime = vue.createTypeScriptRuntime({
+				typescript: ts,
+				vueCompilerOptions,
+				getCssClasses: () => ({}),
+				getCssVBindRanges: () => [],
+				// TODO: move to plugin
+				compileTemplate(template: string, lang: string): {
+					htmlText: string,
+					htmlToTemplate: (start: number, end: number) => { start: number, end: number } | undefined,
+				} | undefined {
+		
+					if (lang === 'html') {
+						return {
+							htmlText: template,
+							htmlToTemplate: (htmlStart, htmlEnd) => ({ start: htmlStart, end: htmlEnd }),
+						};
+					}
+		
+					if (lang === 'pug') {
+		
+						const pugDoc = pug.baseParse(template);
+		
+						if (pugDoc) {
+							return {
+								htmlText: pugDoc.htmlCode,
+								htmlToTemplate: (htmlStart, htmlEnd) => {
+									const pugRange = pugDoc.sourceMap.getSourceRange(htmlStart, htmlEnd, data => !data?.isEmptyTagCompletion)?.[0];
+									if (pugRange) {
+										return pugRange;
+									}
+									else {
+		
+										const pugStart = pugDoc.sourceMap.getSourceRange(htmlStart, htmlStart, data => !data?.isEmptyTagCompletion)?.[0]?.start;
+										const pugEnd = pugDoc.sourceMap.getSourceRange(htmlEnd, htmlEnd, data => !data?.isEmptyTagCompletion)?.[0]?.end;
+		
+										if (pugStart !== undefined && pugEnd !== undefined) {
+											return { start: pugStart, end: pugEnd };
+										}
+									}
+								},
+							};
+						}
+					}
+				},
+				vueHost: proxyHost.host,
+				isTsPlugin: true
+			});
 			const _tsPluginApis = apis.register(tsRuntime.context);
 			const tsPluginProxy: Partial<ts.LanguageService> = {
 				getSemanticDiagnostics: apiHook(tsRuntime.context.scriptTsLsRaw.getSemanticDiagnostics, false),
