+++ packages/vue-typescript/src/index.ts
@@ -3,5 +3,4 @@ export * from './utils/string';
 export * from './vueDocument';
 export * from './vueDocuments';
 export * from './types';
-export * from './basicRuntime';
 export * from './typescriptRuntime';
